requirements.txt
# Core dependencies
opencv-python==4.8.1.78
ultralytics==8.0.196
mediapipe==0.10.7
openvino==2023.1.0

# Database and storage
boto3==1.28.85
python-dotenv==1.0.0

# Natural Language Processing
langchain==0.0.335
llama-cpp-python==0.2.11
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.15

# Web framework and API
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
websockets==12.0

# Notifications
twilio==8.10.0
pushbullet.py==0.12.0

# Audio processing for voice queries
speechrecognition==3.10.0
pyaudio==0.2.11
pyttsx3==2.90

# Utilities
numpy==1.24.3
pillow==10.0.1
requests==2.31.0
python-dateutil==2.8.2
schedule==1.2.0
watchdog==3.0.0

# Development and deployment
docker==6.1.3
pytest==7.4.3
ml-dtypes==0.2.0